requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# Environment
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# HTTP Client (for testing)
httpx==0.26.0

